legacy_code/requirements.txt
requests
yt-dlp
customtkinter
Pillow
tkinterdnd2-universal